move more code into types.rs
